Add reverse color support
In Mathematica there is an option to make background look dark and texts bright, that is to use the "ReverseColor" stylesheet in the View menu. However WolframLanguageForJupyter does not support that, which annoys me as I am gotten used to see dark web pages and that's what I am doing.